azure_ml/dataset/requirements.txt
Faker==22.5.1
feather_format==0.4.1
icecream==2.1.3
pandas==2.2.0
snoop==0.4.3
tqdm==4.64.1
varname==0.12.2
